Separated unit tests and integration tests
