+++ .gitignore
@@ -1,5 +1,6 @@
 *.js
 !jest.config.js
+!jest.integration.config.js
 *.d.ts
 node_modules
