+++ package.json
@@ -5,7 +5,8 @@
     "prioriti-cdk": "bin/prioriti-cdk.js"
   },
   "scripts": {
-    "test": "jest --config=jest.config.js",
+    "test-unit": "jest --config=jest.config.js",
+    "test-integration": "jest --config=jest.integration.config.js",
     "deploy": "cdk deploy --profile personal -c env=prod -c region=eu-west-2",
     "prettier": "prettier --write ./src"
   },
